Change return type of getStudentsQuantityByNames to Long

+++ src/main/java/ait/cohort46/student/service/StudentServiceImpl.java
@@ -72,8 +72,8 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Integer getStudentsQuantityByNames(Set<String> names) {
-        return (int) StreamSupport.stream(studentRepository.findAll().spliterator(), false)
+    public Long getStudentsQuantityByNames(Set<String> names) {
+        return StreamSupport.stream(studentRepository.findAll().spliterator(), false)
                 .filter(s -> names.contains(s.getName()))
                 .count();
     }
